refactor: Updated URLs for Privacy Policy, Terms of Service, and Code of Conduct and updated dependencies
Updated the URLs for the Privacy Policy, Terms of Service, and Code of Conduct across various parts of the application and documentation.

The old Google Sites URLs have been replaced with new GitHub Pages URLs to reflect the new location of these documents.

- Updated links in `preferences_settings.xml`
- Updated links in `StartupActivity.java`
- Updated links in `WebViewActivity.java` and its corresponding raw text file
- Updated links in `HelpActivity.java`
- Updated links in `README.md`

Additionally, the following dependencies and Gradle wrapper were updated:
- `appcompat` to `1.7.1`
- `firebaseBom` to `33.15.0`
- `lifecycle` (livedata, runtime, viewmodel) to `2.9.1`
- `playServicesAds` to `24.4.0`
- Gradle wrapper to `8.14.2`